Add time limit to sub problem solve
When the sub problem takes too long to solve, then add a time limit (and a gap limit) to the problem to escape as soon as possible.

Also changed the logic in making cuts based on solved submodels to only make them when appropriate, and to make appropriate cuts when a larger gap is found.